fix minor error

+++ upbit/upbit_api_trading.py
@@ -45,7 +45,7 @@ def target_price(target):
 def main():
     parser = argparse.ArgumentParser(description="tutorial")
     parser.add_argument('--target-coin', type=str, help='a coin to buy')
-    parser.add_argument('--profit', type=int, help='profit ratio for benefit')
+    parser.add_argument('--profit', type=float, help='profit ratio for benefit')
     args = parser.parse_args()
 
     ############### schedules ###############
